# core/src/main/java/pl/ersms/core/web/request/CreateNewMenuItemRequest.java
package pl.ersms.core.web.request;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.Digits;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

import java.math.BigDecimal;

@JsonIgnoreProperties(ignoreUnknown = true)
public record CreateNewMenuItemRequest(
        @NotBlank(message = "Name cannot be blank")
        @Size(min = 3, max = 50, message = "Name must be between 3 and 50 characters")
        String name,
        @NotBlank(message = "Description cannot be blank")
        @Size(min = 3, max = 4096, message = "Description must be between 3 and 4096 characters")
        String description,
        @DecimalMin(value = "0.0", inclusive = false, message = "Price must be greater than 0")
        @Digits(integer = 10, fraction = 2, message = "Price must be a number with maximum 10 digits and 2 decimal places")
        BigDecimal price
) {
}